Refactor: layouts
- Rename AppLayout to MenuLayout
- Move App related components under ~/pages/apps
